WIP specs

+++ app/controllers/team_members_controller.rb
@@ -4,9 +4,10 @@ class TeamMembersController < ApplicationController
   end
 
   def destroy
-    if org.id != current_user.organisation_id
-    user = User.find(params[:id])
+    user = current_organisation.users.find_by(id: params.fetch(:id))
+    redirect_to team_members_path && return unless user
+
     user.destroy
-    redirect_to team_members_path, notice: "Team member has been successfully removed"
+    redirect_to team_members_path, notice: "Team member has been removed"
   end
 end
